requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning==2.0.8  # >=2.0.0  pip install pytorch-lightning==2.0.8 --upgrade
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550
tqdm
tensorboardX
loguru

# --------- ocsr --------- #
OpenNMT-py==2.2.0
transformers==4.27.1
timm==0.4.12
SmilesPE==0.0.3
albumentations==1.1.0
rdkit==2022.9.5

# --------- render ------- #
shapely
wand  # sudo apt-get install libmagickwand-dev  (sudo apt update)

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
pandas==1.5.3
ipykernel
more_itertools
torch-geometric==2.3.1 
mahotas  # heatmap_max
pyarrow

# --------- plt --------- #
seaborn==0.12.2
cairosvg
bytecode  # attention
nbformat>=4.2.0
plotly==5.18.0